Move everything to a new 'MutedTopics' class
